+++ javaScript/projects/weather-app/src/components/tabs/today/today.config.js
@@ -1,0 +1,32 @@
+export default [
+  {
+    temp_f: 'Temperature',
+  },
+  {
+    feelslike_f: 'Feels like',
+  },
+  {
+    maxtemp_f: 'High',
+  },
+  {
+    mintemp_f: 'Low',
+  },
+  {
+    humidity: 'Humidity',
+  },
+  {
+    pressure_in: 'Pressure',
+  },
+  {
+    vis_miles: 'Visibility',
+  },
+  {
+    wind_mph: 'Wind',
+  },
+  {
+    wind_degree: 'Wind degree',
+  },
+  {
+    wind_dir: 'Wind direction',
+  },
+];